Refactor: Modularize the Debian Package Generator
This commit refactors the monolithic `build_deb_package.py` script into a more modular and maintainable structure.

The key changes include:
- Creating an `app` directory to house the main application logic, making it a Python package.
- Moving the main script logic to `app/main.py`.
- Extracting certificate generation into `app/services/certificate_service.py`.
- Using Jinja2 for templating Debian control files, with templates stored in `debian_templates`.
- Adding a `.gitignore` file to exclude build artifacts and other generated files.
- Updating `requirements.txt` with the `Jinja2` dependency.
- Updating `HOW_TO_RUN.txt` with the new command to run the application.
