+++ .gitignore
@@ -1,0 +1,16 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Build artifacts
+build/
+dist/
+*.deb
+*.tar.gz
+
+# IDE / Editor specific
+.vscode/
+.idea/
+*.swp
+*.swo
