+++ HOW_TO_RUN.txt
@@ -12,7 +12,7 @@ Step 1: Install dependencies
 $ pip3 install -r requirements.txt
 
 Step 2: Run the script
-$ python3 build_deb_package.py
+$ python3 -m app.main
 
 Output: sample-config-pkg_1.0.0_all.deb
 
@@ -58,7 +58,7 @@ TO CREATE EXECUTABLE FROM SCRATCH
 If you want to rebuild the executable:
 
 $ pip3 install -r requirements.txt
-$ pyinstaller --onefile --name generate-deb-package build_deb_package.py
+$ pyinstaller --onefile --name generate-deb-package app/main.py
 
 The executable will be in the dist/ folder.
