+++ app/services/certificate_service.py
@@ -1,0 +1,153 @@
+import os
+from datetime import datetime, timedelta
+from cryptography import x509
+from cryptography.x509.oid import NameOID
+from cryptography.hazmat.primitives import hashes, serialization
+from cryptography.hazmat.primitives.asymmetric import rsa
+from cryptography.hazmat.backends import default_backend
+
+
+def generate_certificates(cert_dir):
+    """
+    Generate CA certificate, server certificate, and client certificate
+    """
+    print("Generating certificates...")
+
+    # Step 1: Generate CA private key and certificate
+    print("  - Generating CA certificate...")
+    ca_key = rsa.generate_private_key(
+        public_exponent=65537,
+        key_size=2048,
+        backend=default_backend()
+    )
+
+    ca_subject = ca_issuer = x509.Name([
+        x509.NameAttribute(NameOID.COUNTRY_NAME, "US"),
+        x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "State"),
+        x509.NameAttribute(NameOID.LOCALITY_NAME, "City"),
+        x509.NameAttribute(NameOID.ORGANIZATION_NAME, "Organization"),
+        x509.NameAttribute(NameOID.COMMON_NAME, "CA"),
+    ])
+
+    ca_cert = x509.CertificateBuilder().subject_name(
+        ca_subject
+    ).issuer_name(
+        ca_issuer
+    ).public_key(
+        ca_key.public_key()
+    ).serial_number(
+        x509.random_serial_number()
+    ).not_valid_before(
+        datetime.utcnow()
+    ).not_valid_after(
+        datetime.utcnow() + timedelta(days=3650)  # 10 years
+    ).add_extension(
+        x509.BasicConstraints(ca=True, path_length=None),
+        critical=True,
+    ).sign(ca_key, hashes.SHA256(), default_backend())
+
+    # Write CA certificate
+    with open(os.path.join(cert_dir, "ca.crt"), "wb") as f:
+        f.write(ca_cert.public_bytes(serialization.Encoding.PEM))
+
+    with open(os.path.join(cert_dir, "ca.key"), "wb") as f:
+        f.write(ca_key.private_bytes(
+            encoding=serialization.Encoding.PEM,
+            format=serialization.PrivateFormat.TraditionalOpenSSL,
+            encryption_algorithm=serialization.NoEncryption()
+        ))
+
+    # Step 2: Generate server private key and certificate (signed by CA)
+    print("  - Generating server certificate...")
+    server_key = rsa.generate_private_key(
+        public_exponent=65537,
+        key_size=2048,
+        backend=default_backend()
+    )
+
+    server_subject = x509.Name([
+        x509.NameAttribute(NameOID.COUNTRY_NAME, "US"),
+        x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "State"),
+        x509.NameAttribute(NameOID.LOCALITY_NAME, "City"),
+        x509.NameAttribute(NameOID.ORGANIZATION_NAME, "Organization"),
+        x509.NameAttribute(NameOID.COMMON_NAME, "server.local"),
+    ])
+
+    server_cert = x509.CertificateBuilder().subject_name(
+        server_subject
+    ).issuer_name(
+        ca_cert.subject
+    ).public_key(
+        server_key.public_key()
+    ).serial_number(
+        x509.random_serial_number()
+    ).not_valid_before(
+        datetime.utcnow()
+    ).not_valid_after(
+        datetime.utcnow() + timedelta(days=365)
+    ).add_extension(
+        x509.SubjectAlternativeName([
+            x509.DNSName("server.local"),
+            x509.DNSName("localhost"),
+        ]),
+        critical=False,
+    ).sign(ca_key, hashes.SHA256(), default_backend())
+
+    # Write server certificate and key
+    with open(os.path.join(cert_dir, "server.crt"), "wb") as f:
+        f.write(server_cert.public_bytes(serialization.Encoding.PEM))
+
+    with open(os.path.join(cert_dir, "server.key"), "wb") as f:
+        f.write(server_key.private_bytes(
+            encoding=serialization.Encoding.PEM,
+            format=serialization.PrivateFormat.TraditionalOpenSSL,
+            encryption_algorithm=serialization.NoEncryption()
+        ))
+
+    # Step 3: Generate client private key and certificate (signed by CA)
+    print("  - Generating client certificate...")
+    client_key = rsa.generate_private_key(
+        public_exponent=65537,
+        key_size=2048,
+        backend=default_backend()
+    )
+
+    client_subject = x509.Name([
+        x509.NameAttribute(NameOID.COUNTRY_NAME, "US"),
+        x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "State"),
+        x509.NameAttribute(NameOID.LOCALITY_NAME, "City"),
+        x509.NameAttribute(NameOID.ORGANIZATION_NAME, "Organization"),
+        x509.NameAttribute(NameOID.COMMON_NAME, "client.local"),
+    ])
+
+    client_cert = x509.CertificateBuilder().subject_name(
+        client_subject
+    ).issuer_name(
+        ca_cert.subject
+    ).public_key(
+        client_key.public_key()
+    ).serial_number(
+        x509.random_serial_number()
+    ).not_valid_before(
+        datetime.utcnow()
+    ).not_valid_after(
+        datetime.utcnow() + timedelta(days=365)
+    ).add_extension(
+        x509.SubjectAlternativeName([
+            x509.DNSName("client.local"),
+        ]),
+        critical=False,
+    ).sign(ca_key, hashes.SHA256(), default_backend())
+
+    # Write client certificate and key
+    with open(os.path.join(cert_dir, "client.crt"), "wb") as f:
+        f.write(client_cert.public_bytes(serialization.Encoding.PEM))
+
+    with open(os.path.join(cert_dir, "client.key"), "wb") as f:
+        f.write(client_key.private_bytes(
+            encoding=serialization.Encoding.PEM,
+            format=serialization.PrivateFormat.TraditionalOpenSSL,
+            encryption_algorithm=serialization.NoEncryption()
+        ))
+
+    print("✓ Certificates generated successfully")
