+++ debian_templates/control.j2
@@ -1,0 +1,9 @@
+Package: {{ package_name }}
+Version: {{ version }}
+Section: utils
+Priority: optional
+Architecture: {{ architecture }}
+Maintainer: {{ maintainer }}
+Description: {{ description }}
+ This package installs certificates (CA, server, client) and configuration.
+ Files are placed under /opt/config/.
