+++ debian_templates/triggers.j2
@@ -1,0 +1,2 @@
+interest-noawait /opt/config/certs
+interest-noawait /opt/config/test.conf
