+++ requirements.txt
@@ -1,2 +1,3 @@
 cryptography>=41.0.0
 pyinstaller>=6.0.0
+Jinja2>=3.1.0
